# sprint4/API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toolshop API",
        "description": "Toolshop REST API technical description",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api-v4.practicesoftwaretesting.com",
            "description": "Deployed environment"
        },
        {
            "url": "http://localhost:8091",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/invoices": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Retrieve all invoices",
                "description": "`user` retrieves only related invoices",
                "operationId": "getInvoices",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagenumber",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "PaginatedInvoiceResponse",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InvoiceResponse"
                                            }
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Store new invoice",
                "description": "Store new invoice",
                "operationId": "storeInvoice",
                "requestBody": {
                    "description": "Invoice request object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntityResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/invoices/{invoiceId}": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Retrieve specific invoice",
                "description": "Retrieve specific invoice",
                "operationId": "getInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "description": "The invoiceId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Update specific invoice",
                "description": "Update specific invoice",
                "operationId": "updateInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "description": "The invoiceId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Invoice request object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/UpdateResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntityResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/invoices/{invoiceId}/status": {
            "put": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Update invoice status",
                "description": "Update invoice status",
                "operationId": "updateInvoiceStatus",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "description": "The invoiceId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "InvoiceStatusRequest",
                                "properties": {
                                    "status": {
                                        "description": "The status of the order",
                                        "type": "string",
                                        "enum": [
                                            "AWAITING_FULFILLMENT",
                                            "ON_HOLD",
                                            "AWAITING_SHIPMENT",
                                            "SHIPPED",
                                            "COMPLETED"
                                        ]
                                    },
                                    "status_message": {
                                        "description": "A message describing the status",
                                        "type": "string",
                                        "maxLength": 50,
                                        "minLength": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/UpdateResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntityResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/invoices/search": {
            "get": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Retrieve specific invoices matching the search query",
                "description": "Search is performed on the `invoice_number`, `billing_address` and `status` column",
                "operationId": "searchInvoice",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "A query phrase",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagenumber",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "PaginatedInvoiceResponse",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InvoiceResponse"
                                            }
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoiceRequest": {
                "title": "InvoiceRequest",
                "properties": {
                    "user_id": {
                        "type": "integer"
                    },
                    "billing_address": {
                        "type": "string"
                    },
                    "billing_city": {
                        "type": "string"
                    },
                    "billing_country": {
                        "type": "string"
                    },
                    "billing_state": {
                        "type": "string"
                    },
                    "billing_postcode": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "Cash on Delivery"
                    },
                    "payment_account_name": {
                        "type": "string",
                        "example": "Jogn Doe"
                    },
                    "payment_account_number": {
                        "type": "string",
                        "example": "0987654345"
                    },
                    "invoicelines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceLineRequest"
                        }
                    }
                },
                "type": "object"
            },
            "InvoiceResponse": {
                "title": "InvoiceResponse",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_date": {
                        "type": "string",
                        "example": "2022-10-20 09:49:45"
                    },
                    "invoice_number": {
                        "type": "string",
                        "example": "INV-2022000002"
                    },
                    "billing_address": {
                        "type": "string"
                    },
                    "billing_city": {
                        "type": "string"
                    },
                    "billing_country": {
                        "type": "string"
                    },
                    "billing_state": {
                        "type": "string"
                    },
                    "billing_postcode": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "Cash on Delivery"
                    },
                    "payment_account_name": {
                        "type": "string",
                        "example": "Jogn Doe"
                    },
                    "payment_account_number": {
                        "type": "string",
                        "example": "0987654345"
                    },
                    "status": {
                        "type": "string",
                        "example": "COMPLETED"
                    },
                    "status_message": {
                        "type": "string",
                        "example": "",
                        "nullable": true
                    },
                    "invoicelines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceLineResponse"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "UpdateResponse": {
                "description": "Result of the update",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "UpdateResponse",
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UnauthorizedResponse": {
                "description": "Returns when user is not authenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "UnauthorizedResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ItemNotFoundResponse": {
                "description": "Returns when the resource is not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "ItemNotFoundResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Requested item not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "MethodNotAllowedResponse": {
                "description": "Returns when the method is not allowed for the requested route",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "MethodNotAllowedResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Method is not allowed for the requested route"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UnprocessableEntityResponse": {
                "description": "Returns when the server was not able to process the content"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}